requirements.txt
black==23.3.0
blinker==1.6.2
click==8.1.3
Flask==2.3.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
mypy-extensions==1.0.0
mysql-connector-python==8.0.33
mysqlclient==2.1.1
packaging==23.1
pathspec==0.11.1
platformdirs==3.4.0
protobuf==3.20.3
PyMySQL==1.0.3
tomli==2.0.1
Werkzeug==2.3.0
